port to riscv64

+++ 12-riscv64/src/syscall.zig
@@ -12,11 +12,11 @@ pub const Sysnum = enum(u8) {
 pub fn call(ctx: *task.TaskRegs) void {
     const sysnum = @intToEnum(Sysnum, ctx.a7);
     switch (sysnum) {
-        .gethartid => ctx.a0 = gethartid(@intToPtr(*align(1) u32, ctx.a0)),
+        .gethartid => ctx.a0 = gethartid(@intToPtr(*align(1) usize, ctx.a0)),
     }
 }
 
-fn gethartid(id: *align(1) u32) u32 {
+fn gethartid(id: *align(1) usize) usize {
     print("--> sys_gethid, arg0 = 0x{x}\n", .{@ptrToInt(id)}); // the alignment of ctx.a0 is not 4
     id.* = csr.read("mhartid");
     return 0;
